requirements.txt
fastapi
uvicorn
pydantic
pandas
joblib
openai
neo4j
python-dotenv
requests
scikit-learn
pymupdf